+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### BREAKING CHANGES
 - Added eslint-plugin-jest to peerDependencies.
+- Added eslint-plugin-unicorn to peerDependencies.
 - Changed rule 'tslint/object-literal-sort-keys' to 'sort-keys'  
 - Dropped support for Node.js 10.
 
@@ -15,6 +16,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support for eslint-plugin-sonarjs@^0.8.0
 - Support for eslint-plugin-sonarjs@^0.9.0
 
+### Removed
+- @typescript-eslint/eslint-plugin-tslint
+- tslint
+ 
 ## [3.1.0] - 2021-05-05
 ### Added
 - Support for eslint-plugin-sonarjs@^0.6.0